requirements.txt
AWSIoTPythonSDK==1.5.4
boto3==1.34.75
botocore==1.34.75
colorzero==2.0
gpiozero==2.0.1
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
